+++ CHANGELOG.md
@@ -2,8 +2,21 @@
 
 All notable changes to the "sfk-studio" extension will be documented in this file.
 
-Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
-
 ## [Unreleased]
 
-- Initial release
+- [GoToDefinition] : Implementation for scripts
+
+## [1.0.1] - 2020-03-20
+
+### Added
+
+- [GoToDefinition] : Support for DataSourceDefaultValue property
+- [GoToDefinition] : Search file in the current DeviceType otherwise in Common
+
+## [1.0.0] - 2020-03-13
+
+- [GoToDefinition]: Initial feature
+
+[Unreleased]: https://projects-dgs.divalto.com/weavy/tools/_git/vscode.sfk-studio?version=GBdevelop
+[1.0.1]: https://projects-dgs.divalto.com/weavy/tools/_git/vscode.sfk-studio?version=GT1.0.1
+[1.0.0]: https://projects-dgs.divalto.com/weavy/tools/_git/vscode.sfk-studio?version=GCb6107068e738855454b2f4e681af6716a84c5f59